Extract Prompt Studio extend slot controls
